requirements.txt
streamlit
plotly
pandas
networkx
python-dotenv
openai

